refactor: add options to hide schedule
