+++ packages/utils/src/date-utils/conversionUtils.test.ts
@@ -1,4 +1,4 @@
-import { millisToHours, millisToMinutes, millisToSeconds } from "./conversionUtils";
+import { millisToHours, millisToMinutes, millisToSeconds, secondsInMillis } from './conversionUtils';
 
 describe('millisToSecond()', () => {
   test('null values', () => {
@@ -110,5 +110,16 @@ describe('millisToHours()', () => {
     // negative numbers are rounded up
     const t = { val: -86401000, result: -25 };
     expect(millisToHours(t.val)).toBe(t.result);
+  });
+});
+
+describe('secondsInMillis()', () => {
+  it('return 0 if value is null', () => {
+    expect(secondsInMillis(null)).toBe(0);
+  });
+  it('returns the seconds value of a millis date', () => {
+    const date = 1686255053619; // Thu Jun 08 2023 20:10:53
+    const seconds = secondsInMillis(date);
+    expect(seconds).toBe(53);
   });
 });
